Initial commit.
NOTE: this commit reflects the state of the code 'as was' at the outset of the challenge. It won't work as is. One reason is that required functions have changed or gone. Second, the scripts in folder scripts/data, originally located in the raw data directory, were moved here without adapting the paths. This as well as other fixes are planned for the following commits.